feature: add kahoot logic

+++ domain/src/main/java/eu/batomobile/domain/entity/Question.kt
@@ -4,7 +4,7 @@ package eu.batomobile.domain.entity
 data class Question(
     var type: String? = null,
     var question: String? = null,
-    var time: Int? = null,
+    var time: Long? = null,
     var points: Boolean? = null,
     var pointsMultiplier: Int? = null,
     var choices: ArrayList<Choice> = arrayListOf(),
